chore: move useLocalStorage to hooks
