+++ src/pages/event/EventPage.jsx
@@ -8,7 +8,7 @@ import { Splitter, SplitterPanel } from 'primereact/splitter'
 import EventList from './EventList'
 import useSearchFilter from '/src/hooks/useSearchFilter'
 import { toast } from 'react-toastify'
-import { useLocalStorage } from '/src/utils'
+import useLocalStorage from '/src/hooks/useLocalStorage'
 import EventOverview from './EventOverview'
 import { StringParam, useQueryParam } from 'use-query-params'
 import { useMemo } from 'react'
