+++ src/utils.js
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import { isEmpty, isEqual, xorWith, cloneDeep, sortBy } from 'lodash'
 
 const arrayEquals = (x, y) => isEmpty(xorWith(x, y, isEqual))
@@ -172,44 +171,6 @@ const getTaskTypes = () => {
   return sortBy(result, 'name')
 }
 
-function useLocalStorage(key, initialValue) {
-  // State to store our value
-  // Pass initial state function to useState so logic is only executed once
-  const [storedValue, setStoredValue] = useState(() => {
-    if (typeof window === 'undefined') {
-      return initialValue
-    }
-    try {
-      // Get from local storage by key
-      const item = window.localStorage.getItem(key)
-      // Parse stored json or if none return initialValue
-      return item ? JSON.parse(item) : initialValue
-    } catch (error) {
-      // If error also return initialValue
-      console.log(error)
-      return initialValue
-    }
-  })
-  // Return a wrapped version of useState's setter function that ...
-  // ... persists the new value to localStorage.
-  const setValue = (value) => {
-    try {
-      // Allow value to be a function so we have same API as useState
-      const valueToStore = value instanceof Function ? value(storedValue) : value
-      // Save state
-      setStoredValue(valueToStore)
-      // Save to local storage
-      if (typeof window !== 'undefined') {
-        window.localStorage.setItem(key, JSON.stringify(valueToStore))
-      }
-    } catch (error) {
-      // A more advanced implementation would handle the error case
-      console.log(error)
-    }
-  }
-  return [storedValue, setValue]
-}
-
 //
 // Export
 //
@@ -232,5 +193,4 @@ export {
   updateStatusIcons,
   updateStatusShortNames,
   updateTagColors,
-  useLocalStorage,
 }
